+++ tests/test_subversion.py
@@ -114,6 +114,29 @@ class SubversionTestCase(SCMVerTestCase):
                 self.switch(path)
                 self.assertEqual(svn.parse('.', name='.svn', **kwargs), core.SCMInfo(tag, 0, 3, False, branch))
 
+    def test_i18n(self):
+        self.check_locale()
+
+        self.create('repo')
+        self.checkout('repo', 'wc')
+        svn.run('mkdir', 'trunk', 'branches', 'tags')
+        svn.run('commit', '-m', '_')
+        svn.run('copy', 'trunk', os.path.join('branches', u'\u30d6\u30e9\u30f3\u30c1'))
+        svn.run('commit', '-m', '_')
+        svn.run('copy', os.path.join('branches', u'\u30d6\u30e9\u30f3\u30c1'), os.path.join('tags', u'\u30bf\u30b0'))
+        svn.run('commit', '-m', '_')
+
+        for path, branch in (
+            ('', None),
+            ('trunk', 'trunk'),
+            ('branches', None),
+            (u'branches/\u30d6\u30e9\u30f3\u30c1', u'\u30d6\u30e9\u30f3\u30c1'),
+            ('tags', None),
+            (u'tags/\u30bf\u30b0', None),
+        ):
+            self.switch(path)
+            self.assertEqual(svn.parse('.', name='.svn'), core.SCMInfo(u'\u30bf\u30b0', 0, 3, False, branch))
+
     def test_status(self):
         self.create('repo')
         self.checkout('repo', 'wc')